+++ NEWS.md
@@ -1,4 +1,5 @@
 # fdasrvf (development version)
+* add `interparc` function for downsampling closed curves
 
 # fdasrvf 2.4.4
 * expose PNS functions